refactor: cleanup and remove MediaIcon
