+++ packages/shared/components/molecules/index.js
@@ -6,12 +6,11 @@ export { default as BasicActivityDetails } from './BasicActivityDetails.svelte'
 export { default as DateInputButton } from './DateInputButton.svelte'
 export { default as ExpirationTimePicker } from './ExpirationTimePicker.svelte'
 export { default as FoundryActivityDetails } from './FoundryActivityDetails.svelte'
-export { default as MediaIcon } from './MediaIcon.svelte'
 export { default as MediaPlaceholder } from './MediaPlaceholder.svelte'
 export { default as NftActivityDetails } from './NftActivityDetails.svelte'
 export { default as NftGallery } from './NftGallery.svelte'
 export { default as NftGalleryItem } from './NftGalleryItem.svelte'
-export { default as NftImageOrIcon } from './NftImageOrIcon.svelte'
+export { default as NftImageOrIconBox } from './NftImageOrIconBox.svelte'
 export { default as NftMedia } from './NftMedia.svelte'
 export { default as SelectorInput } from './SelectorInput.svelte'
 export { default as ShimmerClaimingAccountList } from './ShimmerClaimingAccountList.svelte'
